requirements.txt
praw>=7.7.0
pyyaml>=6.0
python-dotenv>=1.0.0
playwright>=1.40.0
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.1
beautifulsoup4>=4.12.0
requests>=2.31.0
